# Conflicts:
#	README.md
#	build/cmake/gameanalytics/CMakeLists.txt
#	build/jenkins/build.py
#	build/jenkins/install_dependencies.py
#	precompiled/linux-x64-gcc-shared/Release/libGameAnalytics.so
#	precompiled/linux-x64-gcc-static/Debug/libGameAnalytics.a
#	precompiled/linux-x64-gcc-static/Release/libGameAnalytics.a
#	precompiled/linux-x86-gcc-shared/Debug/libGameAnalytics.so
#	precompiled/linux-x86-gcc-shared/Release/libGameAnalytics.so
#	precompiled/linux-x86-gcc-static/Debug/libGameAnalytics.a
#	precompiled/linux-x86-gcc-static/Release/libGameAnalytics.a
#	precompiled/osx-shared/Debug/libGameAnalytics.dylib
#	precompiled/osx-shared/Release/GameAnalytics.bundle
#	precompiled/osx-shared/Release/libGameAnalytics.dylib
#	precompiled/osx-static/Debug/libGameAnalytics.a
#	precompiled/osx-static/Release/libGameAnalytics.a
#	precompiled/tizen-arm-shared/Debug/libGameAnalytics.so
#	precompiled/tizen-arm-shared/Release/libGameAnalytics.so
#	precompiled/tizen-arm-static/Debug/libGameAnalytics.a
#	precompiled/tizen-arm-static/Release/libGameAnalytics.a
#	precompiled/tizen-x86-shared/Debug/libGameAnalytics.so
#	precompiled/tizen-x86-shared/Release/libGameAnalytics.so
#	precompiled/tizen-x86-static/Debug/libGameAnalytics.a
#	precompiled/tizen-x86-static/Release/libGameAnalytics.a
#	precompiled/uwp-arm-vc140-shared/Debug/GameAnalytics.dll
#	precompiled/uwp-arm-vc140-shared/Debug/GameAnalytics.exp
#	precompiled/uwp-arm-vc140-shared/Debug/GameAnalytics.lib
#	precompiled/uwp-arm-vc140-shared/Debug/GameAnalytics.pdb
#	precompiled/uwp-arm-vc140-shared/Release/GameAnalytics.dll
#	precompiled/uwp-arm-vc140-shared/Release/GameAnalytics.exp
#	precompiled/uwp-arm-vc140-shared/Release/GameAnalytics.lib
#	precompiled/uwp-arm-vc140-static/Debug/GameAnalytics.lib
#	precompiled/uwp-arm-vc140-static/Debug/GameAnalytics.pdb
#	precompiled/uwp-arm-vc140-static/Release/GameAnalytics.lib
#	precompiled/uwp-x64-vc140-shared/Debug/GameAnalytics.dll
#	precompiled/uwp-x64-vc140-shared/Debug/GameAnalytics.exp
#	precompiled/uwp-x64-vc140-shared/Debug/GameAnalytics.lib
#	precompiled/uwp-x64-vc140-shared/Debug/GameAnalytics.pdb
#	precompiled/uwp-x64-vc140-shared/Release/GameAnalytics.dll
#	precompiled/uwp-x64-vc140-shared/Release/GameAnalytics.exp
#	precompiled/uwp-x64-vc140-shared/Release/GameAnalytics.lib
#	precompiled/uwp-x64-vc140-static/Debug/GameAnalytics.lib
#	precompiled/uwp-x64-vc140-static/Debug/GameAnalytics.pdb
#	precompiled/uwp-x64-vc140-static/Release/GameAnalytics.lib
#	precompiled/uwp-x86-vc140-shared/Debug/GameAnalytics.dll
#	precompiled/uwp-x86-vc140-shared/Debug/GameAnalytics.exp
#	precompiled/uwp-x86-vc140-shared/Debug/GameAnalytics.lib
#	precompiled/uwp-x86-vc140-shared/Debug/GameAnalytics.pdb
#	precompiled/uwp-x86-vc140-shared/Release/GameAnalytics.dll
#	precompiled/uwp-x86-vc140-shared/Release/GameAnalytics.exp
#	precompiled/uwp-x86-vc140-shared/Release/GameAnalytics.lib
#	precompiled/uwp-x86-vc140-static/Debug/GameAnalytics.lib
#	precompiled/uwp-x86-vc140-static/Debug/GameAnalytics.pdb
#	precompiled/uwp-x86-vc140-static/Release/GameAnalytics.lib
#	precompiled/win32-vc120-mt-static/Debug/GameAnalytics.lib
#	precompiled/win32-vc120-mt-static/Release/GameAnalytics.lib
#	precompiled/win32-vc120-shared/Debug/GameAnalytics.dll
#	precompiled/win32-vc120-shared/Debug/GameAnalytics.exp
#	precompiled/win32-vc120-shared/Debug/GameAnalytics.ilk
#	precompiled/win32-vc120-shared/Debug/GameAnalytics.lib
#	precompiled/win32-vc120-shared/Debug/GameAnalytics.pdb
#	precompiled/win32-vc120-shared/Release/GameAnalytics.dll
#	precompiled/win32-vc120-shared/Release/GameAnalytics.exp
#	precompiled/win32-vc120-shared/Release/GameAnalytics.lib
#	precompiled/win32-vc120-static/Debug/GameAnalytics.lib
#	precompiled/win32-vc120-static/Release/GameAnalytics.lib
#	precompiled/win32-vc140-mt-static/Debug/GameAnalytics.lib
#	precompiled/win32-vc140-mt-static/Debug/GameAnalytics.pdb
#	precompiled/win32-vc140-mt-static/Release/GameAnalytics.lib
#	precompiled/win32-vc140-shared-nowmi/Debug/GameAnalytics.dll
#	precompiled/win32-vc140-shared-nowmi/Debug/GameAnalytics.exp
#	precompiled/win32-vc140-shared-nowmi/Debug/GameAnalytics.ilk
#	precompiled/win32-vc140-shared-nowmi/Debug/GameAnalytics.lib
#	precompiled/win32-vc140-shared-nowmi/Debug/GameAnalytics.pdb
#	precompiled/win32-vc140-shared-nowmi/Release/GameAnalytics.dll
#	precompiled/win32-vc140-shared-nowmi/Release/GameAnalytics.exp
#	precompiled/win32-vc140-shared-nowmi/Release/GameAnalytics.lib
#	precompiled/win32-vc140-shared/Debug/GameAnalytics.dll
#	precompiled/win32-vc140-shared/Debug/GameAnalytics.exp
#	precompiled/win32-vc140-shared/Debug/GameAnalytics.ilk
#	precompiled/win32-vc140-shared/Debug/GameAnalytics.lib
#	precompiled/win32-vc140-shared/Debug/GameAnalytics.pdb
#	precompiled/win32-vc140-shared/Release/GameAnalytics.dll
#	precompiled/win32-vc140-shared/Release/GameAnalytics.exp
#	precompiled/win32-vc140-shared/Release/GameAnalytics.lib
#	precompiled/win32-vc140-static/Debug/GameAnalytics.lib
#	precompiled/win32-vc140-static/Debug/GameAnalytics.pdb
#	precompiled/win32-vc140-static/Release/GameAnalytics.lib
#	precompiled/win64-vc120-mt-static/Debug/GameAnalytics.lib
#	precompiled/win64-vc120-mt-static/Release/GameAnalytics.lib
#	precompiled/win64-vc120-shared/Debug/GameAnalytics.dll
#	precompiled/win64-vc120-shared/Debug/GameAnalytics.exp
#	precompiled/win64-vc120-shared/Debug/GameAnalytics.ilk
#	precompiled/win64-vc120-shared/Debug/GameAnalytics.lib
#	precompiled/win64-vc120-shared/Debug/GameAnalytics.pdb
#	precompiled/win64-vc120-shared/Release/GameAnalytics.dll
#	precompiled/win64-vc120-shared/Release/GameAnalytics.exp
#	precompiled/win64-vc120-shared/Release/GameAnalytics.lib
#	precompiled/win64-vc120-static/Debug/GameAnalytics.lib
#	precompiled/win64-vc120-static/Release/GameAnalytics.lib
#	precompiled/win64-vc140-mt-static/Debug/GameAnalytics.lib
#	precompiled/win64-vc140-mt-static/Debug/GameAnalytics.pdb
#	precompiled/win64-vc140-mt-static/Release/GameAnalytics.lib
#	precompiled/win64-vc140-shared/Debug/GameAnalytics.dll
#	precompiled/win64-vc140-shared/Debug/GameAnalytics.exp
#	precompiled/win64-vc140-shared/Debug/GameAnalytics.ilk
#	precompiled/win64-vc140-shared/Debug/GameAnalytics.lib
#	precompiled/win64-vc140-shared/Debug/GameAnalytics.pdb
#	precompiled/win64-vc140-shared/Release/GameAnalytics.dll
#	precompiled/win64-vc140-shared/Release/GameAnalytics.exp
#	precompiled/win64-vc140-shared/Release/GameAnalytics.lib
#	precompiled/win64-vc140-static/Debug/GameAnalytics.lib
#	precompiled/win64-vc140-static/Debug/GameAnalytics.pdb
#	precompiled/win64-vc140-static/Release/GameAnalytics.lib
#	source/dependencies/crypto/hmac_sha2.c
#	source/dependencies/crypto/hmac_sha2.h
#	source/dependencies/curl/build.sh
#	source/dependencies/curl/lib/osx/libcurl.a
#	source/dependencies/miniz/miniz.c
#	source/dependencies/openssl/1.0.2h/build.sh
#	source/gameanalytics/Foundation/GAThreadHelpers.h
#	source/gameanalytics/GADevice.cpp
#	source/gameanalytics/GAEvents.cpp
#	source/gameanalytics/GALogger.cpp
#	source/gameanalytics/GAState.h
#	source/gameanalytics/GAThreading.cpp
#	source/gameanalytics/GAThreading.h
#	source/gameanalytics/GAValidator.cpp
#	source/gameanalytics/GameAnalytics.cpp
#	source/gameanalytics/GameAnalytics.h
#	source/gameanalytics/GameAnalyticsExtern.cpp
#	source/gameanalytics/GameAnalyticsExtern.h

+++ tests/run_tests_osx.py
@@ -2,10 +2,12 @@
 import os
 import subprocess
 import shutil
-
+from os.path import expanduser
 
 BUILD_ROOT = os.path.abspath(os.path.join(__file__, '..'))
+CMAKE_BIN_FILE = os.path.abspath(os.path.join(__file__, '..', '..', 'build', 'jenkins', 'dependencies', 'cmake', 'bin', 'cmake'))
 BUILD_DIR = os.path.join(BUILD_ROOT, "build")
+GA_DIR = os.path.join(expanduser("~"), "GameAnalytics")
 
 
 def make_test_build_dir():
@@ -14,15 +16,23 @@ def make_test_build_dir():
     except OSError:
         pass
 
+
+def remove_gameanalytics_dir():
+    if os.path.exists(GA_DIR):
+        shutil.rmtree(GA_DIR)
+
+
 def remove_test_build_dir():
-    shutil.rmtree(BUILD_DIR)
+    if os.path.exists(BUILD_DIR):
+        shutil.rmtree(BUILD_DIR)
+
 
 def change_to_build_dir():
     os.chdir(BUILD_DIR)
 
 
 def run_cmake_tests():
-    cmake_call = 'cmake -DPLATFORM:STRING=osx-static -G Xcode ..'
+    cmake_call = CMAKE_BIN_FILE + ' -DPLATFORM:STRING=osx-static -G Xcode ..'
     subprocess.call(cmake_call, shell=True)
 
 
@@ -32,11 +42,13 @@ def compile_and_run_tests():
 
 
 def main():
+    remove_gameanalytics_dir()
     remove_test_build_dir()
     make_test_build_dir()
     change_to_build_dir()
     run_cmake_tests()
     compile_and_run_tests()
+
 
 if __name__ == '__main__':
     main()
